save 0 norm

+++ src/handle_cspn.c
@@ -6,7 +6,7 @@
 /*   By: pgerbaud <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/08/19 14:50:28 by pgerbaud          #+#    #+#             */
-/*   Updated: 2018/01/04 20:19:41 by pgerbaud         ###   ########.fr       */
+/*   Updated: 2018/01/05 13:46:59 by pgerbaud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,6 @@ char		*handle_s(t_conv **conv, char *fmt)
 	char	*rslt;
 	int		max;
 
-	//printf("HANDLE_S 0 fmt.%p\n", fmt);
 	max = (*conv)->precision;
 	str = (*conv)->sdata;
 	rslt = NULL;
@@ -46,21 +45,15 @@ char		*handle_s(t_conv **conv, char *fmt)
 		rslt = wcharstr_to_str(str, conv);
 	else
 		rslt = ft_strdup((char *)(*conv)->sdata);
-	if (ft_strlen(rslt) > (*conv)->precision && (*conv)->prec_changed)
+	if ((int)ft_strlen(rslt) > (*conv)->precision && (*conv)->prec_changed)
 		*(rslt + (*conv)->precision) = '\0';
-	//printf("HANDLE_S 1 fmt.%p _%s_ %d\n", rslt, rslt, ft_strlen(rslt));
 	if (ft_strlen(rslt) == 0)
 	{
-		//rslt = ft_strnew(2);
 		*fmt = 0;
 		*(fmt + 1) = '\0';
-		//*rslt = (char)0;
-		//*(rslt + 1) = '\0';
-		//printf("HANDLE_S 1.0 fmt.%p\n", fmt);
 	}
 	else
 		fmt = ft_addinstr(fmt, rslt, "%", 0);
-	//printf("HANDLE_S 2 rslt.%p %d\n", rslt, ft_strlen(rslt));
 	ft_strdel(&rslt);
 	return (fmt);
 }
